third_party/WebKit/LayoutTests/http/tests/inspector/search/source-frame-search-expected.txt

Tests different types of search in SourceFrame


Running: testSearch
Running search test for query = searchTestUniqueString:
  - function <searchTestUniqueString>()
  -     // <searchTestUniqueString> two occurences on the same line searchTestUniqueString
  -     // searchTestUniqueString two occurences on the same line <searchTestUniqueString>
  -     // <searchTestUniqueString> on the next line.
  -     <searchTestUniqueString>();
  -     // <SEARCHTestUniqueString>();

Running: testSearchCaseSensitive
Running search test for query = SEARCHTestUniqueString (caseSensitive):
  -     // <SEARCHTestUniqueString>();

Running: testSearchRegex
Running search test for query = searchTestUnique.* (regex):
  - function <searchTestUniqueString()>
  -     // <searchTestUniqueString two occurences on the same line searchTestUniqueString>
  -     // <searchTestUniqueString on the next line.>
  -     <searchTestUniqueString();>
  -     // <SEARCHTestUniqueString();>
  - // <searchTestUnique space String>

Running: testSearchCaseSensitiveRegex
Running search test for query = searchTestUnique.* (regex, caseSensitive):
  - function <searchTestUniqueString()>
  -     // <searchTestUniqueString two occurences on the same line searchTestUniqueString>
  -     // <searchTestUniqueString on the next line.>
  -     <searchTestUniqueString();>
  - // <searchTestUnique space String>

Running: testSearchConsequent
Running search test for query = AAAAA:
  - // <AAAAA>AAAAAA 11xA here
  - // AAAAA<AAAAA>A 11xA here

Running: testReplace
Running replace test for /REPLACEME1/REPLACED/  (caseSensitive):
  - // REPLACE#ME1#D#
Running replace test for /REPLACEME2/REPLACED/  (caseSensitive, replaceAll):
  - // REPLACE#ME2 REPLACEME2#D REPLACED#
Running replace test for /replaceMe1/replaced/ :
  - // replace#Me1#d#
Running replace test for /replaceMe2/replaced/  (replaceAll):
  - // replace#Me2 replaceMe2#d replaced#
Running replace test for /(REPLACE)ME[38]/$1D/  (regex, caseSensitive):
  - // REPLACE#ME3#D#
Running replace test for /(REPLACE)ME[45]/$1D/  (regex, caseSensitive, replaceAll):
  - // REPLACE#ME4 REPLACEME4#D REPLACED#
Running replace test for /(replac)eMe[38]/$1d/  (regex):
  - // replac#eMe3#d#
Running replace test for /(replace)Me[45]/$1d/  (regex, replaceAll):
  - // replace#Me4 replaceMe4#d replaced#
